Extract loading / dealing with data into NoteInformationController
