+++ Simplenote/Information/NoteInformationViewController.swift
@@ -3,24 +3,33 @@ import SimplenoteFoundation
 
 // MARK: - NoteInformationViewController
 //
-class NoteInformationViewController: UIViewController {
+final class NoteInformationViewController: UIViewController {
     @IBOutlet private weak var tableView: UITableView!
     @IBOutlet private weak var screenTitleLabel: UILabel!
     @IBOutlet private weak var dismissButton: UIButton!
 
     private var transitioningManager: UIViewControllerTransitioningDelegate?
 
-    private let note: Note
-    private var rows: [Row] = []
+    private var rows: [NoteInformationController.Row] = []
+    private let controller: NoteInformationController
 
     /// Designated initializer
     ///
     /// - Parameters:
+    ///     - controller: NoteInformationController
+    ///
+    init(controller: NoteInformationController) {
+        self.controller = controller
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    /// Convenience initializer
+    ///
+    /// - Parameters:
     ///     - note: Note
     ///
-    init(note: Note) {
-        self.note = note
-        super.init(nibName: nil, bundle: nil)
+    convenience init(note: Note) {
+        self.init(controller: NoteInformationController(note: note))
     }
 
     required init?(coder: NSCoder) {
@@ -38,33 +47,21 @@ class NoteInformationViewController: UIViewController {
         configureAccessibility()
 
         startListeningToNotifications()
-
-        reloadData()
+        startListeningForControllerChanges()
     }
 }
 
-// MARK: - Data
+// MARK: - Controller
 private extension NoteInformationViewController {
-    func reloadData() {
-        rows = metricRows()
-        tableView.reloadData()
+    func startListeningForControllerChanges() {
+        controller.observer = { [weak self] rows in
+            self?.update(with: rows)
+        }
     }
 
-    func metricRows() -> [Row] {
-        let metrics = NoteMetrics(note: note)
-        return [
-            .metric(title: Localization.modified,
-                    value: DateFormatter.dateTimeFormatter.string(from: metrics.modifiedDate)),
-
-            .metric(title: Localization.created,
-                    value: DateFormatter.dateTimeFormatter.string(from: metrics.creationDate)),
-
-            .metric(title: Localization.words,
-                    value: NumberFormatter.decimalFormatter.string(for: metrics.numberOfWords)),
-
-            .metric(title: Localization.characters,
-                    value: NumberFormatter.decimalFormatter.string(for: metrics.numberOfChars))
-        ]
+    func update(with rows: [NoteInformationController.Row]) {
+        self.rows = rows
+        tableView.reloadData()
     }
 }
 
@@ -197,16 +194,7 @@ extension NoteInformationViewController {
     }
 }
 
-private enum Row {
-    case metric(title: String, value: String?)
-}
-
 private struct Localization {
     static let information = NSLocalizedString("Information", comment: "Card title showing information about the note (metrics, references)")
-    static let modified = NSLocalizedString("Modified", comment: "Note Modification Date")
-    static let created = NSLocalizedString("Created", comment: "Note Creation Date")
-    static let words = NSLocalizedString("Words", comment: "Number of words in the note")
-    static let characters = NSLocalizedString("Characters", comment: "Number of characters in the note")
-
     static let dismissAccessibilityLabel = NSLocalizedString("Dismiss Information", comment: "Accessibility label describing a button used to dismiss an information view of the note")
 }
